input on own page
